Add PascalJsonPropertyNameTransform and refactor to use NoSeparatorJsonPropertyNameTransform
